requirements.txt
# FastAPI 核心框架
fastapi==0.115.6
uvicorn[standard]==0.34.0

# 数据库相关
sqlmodel==0.0.22

# 身份验证和安全
passlib==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.20

# 文件处理
aiofiles==24.1.0
